+++ sites/www/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`1076` Delay between starting parallel processes
 * :release:`1.10.2 <2015-06-19>`
 * :support:`1325` Clarify `~fabric.operations.put` docs re: the ``mode``
   argument. Thanks to ``@mjmare`` for the catch.